requirements.txt
httpx
openpyxl
pandas
pypdf<5
python-dateutil
python-dotenv
PyYAML>=6
SQLAlchemy>=2.0,<3
fastapi>=0.110
uvicorn>=0.27
